+++ src/animation.c
@@ -1746,7 +1746,7 @@ enum autodraw_types* autodraw_ptr;
 
             if (otmp->quan > 1)
             {
-                int glyph_idx = (otmp->quan <= 8 ? otmp->quan - 2 : otmp->quan <= 30 ? 7 : otmp->quan <= 100 ? 8 : otmp->quan <= 1000 ? 9 : 10);
+                int glyph_idx = (otmp->quan <= 8 ? (int)otmp->quan - 2 : otmp->quan <= 30 ? 7 : otmp->quan <= 100 ? 8 : otmp->quan <= 1000 ? 9 : 10);
                 if (autodraw_ptr)
                     *autodraw_ptr = replacements[replacement_idx].tile_autodraw[glyph_idx];
                 return glyph2tile[glyph_idx + replacement_offsets[replacement_idx] /* replacements[replacement_idx].glyph_offset */ + GLYPH_REPLACEMENT_OFF];
@@ -1766,7 +1766,7 @@ enum autodraw_types* autodraw_ptr;
 
             if (otmp->quan > 1)
             {
-                int glyph_idx = (otmp->quan <= 8 ? otmp->quan - 2 : otmp->quan <= 20 ? 7 : otmp->quan <= 50 ? 8 : otmp->quan <= 100 ? 9 : 10);
+                int glyph_idx = (otmp->quan <= 8 ? (int)otmp->quan - 2 : otmp->quan <= 20 ? 7 : otmp->quan <= 50 ? 8 : otmp->quan <= 100 ? 9 : 10);
                 if (autodraw_ptr)
                     *autodraw_ptr = replacements[replacement_idx].tile_autodraw[glyph_idx];
                 return glyph2tile[glyph_idx + replacement_offsets[replacement_idx] /* replacements[replacement_idx].glyph_offset */ + GLYPH_REPLACEMENT_OFF];
@@ -2327,7 +2327,7 @@ struct replacement_info info;
 
             if (otmp->quan > 1)
             {
-                int glyph_idx = (otmp->quan <= 8 ? otmp->quan - 2 : otmp->quan <= 30 ? 7 : otmp->quan <= 100 ? 8 : otmp->quan <= 1000 ? 9 : 10);
+                int glyph_idx = (otmp->quan <= 8 ? (int)otmp->quan - 2 : otmp->quan <= 30 ? 7 : otmp->quan <= 100 ? 8 : otmp->quan <= 1000 ? 9 : 10);
                 return sign * (glyph_idx + replacement_offsets[replacement_idx] /* replacements[replacement_idx].glyph_offset */ + GLYPH_REPLACEMENT_OFF);
             }
             break;
@@ -2342,7 +2342,7 @@ struct replacement_info info;
 
             if (otmp->quan > 1)
             {
-                int glyph_idx = (otmp->quan <= 8 ? otmp->quan - 2 : otmp->quan <= 20 ? 7 : otmp->quan <= 50 ? 8 : otmp->quan <= 100 ? 9 : 10);
+                int glyph_idx = (otmp->quan <= 8 ? (int)otmp->quan - 2 : otmp->quan <= 20 ? 7 : otmp->quan <= 50 ? 8 : otmp->quan <= 100 ? 9 : 10);
                 return sign * (glyph_idx + replacement_offsets[replacement_idx] /* replacements[replacement_idx].glyph_offset */ + GLYPH_REPLACEMENT_OFF);
             }
             break;
@@ -3509,7 +3509,7 @@ int spef_number;
 {
     if (context.spef_intervals_to_wait_until_action[spef_number] > 0UL)
     {
-        delay_output_intervals(context.spef_intervals_to_wait_until_action[spef_number]);
+        delay_output_intervals((int)context.spef_intervals_to_wait_until_action[spef_number]);
         context.spef_intervals_to_wait_until_action[spef_number] = 0UL;
     }
 }
@@ -3520,7 +3520,7 @@ int spef_number;
 {
     if (context.spef_intervals_to_wait_until_end[spef_number] > 0)
     {
-        delay_output_intervals(context.spef_intervals_to_wait_until_end[spef_number]);
+        delay_output_intervals((int)context.spef_intervals_to_wait_until_end[spef_number]);
         context.spef_intervals_to_wait_until_end[spef_number] = 0UL;
     }
 
@@ -3545,8 +3545,8 @@ int milliseconds;
     if (milliseconds < 0)
         return;
 
-    unsigned long interval_length = (flags.animation_frame_interval_in_milliseconds > 0 ? flags.animation_frame_interval_in_milliseconds : ANIMATION_FRAME_INTERVAL);
-    unsigned long intervals = (unsigned long)milliseconds / interval_length;
+    int interval_length = (flags.animation_frame_interval_in_milliseconds > 0 ? flags.animation_frame_interval_in_milliseconds : ANIMATION_FRAME_INTERVAL);
+    int intervals = milliseconds / interval_length;
 
     reduce_counters_intervals(intervals);
 }
